Add files via upload
Can navigate through directories and list the contents of directories remotely from the client. in order to connect to the victim the IP address must be known and must be changed at the top of the client program. Error checking needs to be done, as I've only tested program with one file at a time.